+++ src/navigation/AppNavigator/AppNavigator.routes.ts
@@ -1,12 +1,12 @@
-import { CardNavigatorRouteParamsList } from '@navigation/CardNavigator/CardNavigator.routes';
-import { NavigatorScreenParams } from '@react-navigation/native';
+// import { CardNavigatorRouteParamsList } from '@navigation/CardNavigator/CardNavigator.routes';
+// import { NavigatorScreenParams } from '@react-navigation/native';
 
-export enum AppNavigatorRouteNames {
-  Home = 'Home',
-  CardNavigator = 'CardNavigator',
-}
+// export enum AppNavigatorRouteNames {
+//   Home = 'Home',
+//   CardNavigator = 'CardNavigator',
+// }
 
-export type AppNavigatorRouteParamsList = {
-  Home: undefined;
-  CardNavigator: NavigatorScreenParams<CardNavigatorRouteParamsList>;
-};
+// export type AppNavigatorRouteParamsList = {
+//   Home: undefined;
+//   CardNavigator: NavigatorScreenParams<CardNavigatorRouteParamsList>;
+// };
