+++ tsconfig.json
@@ -7,6 +7,15 @@
     "noEmit": true,
     "skipLibCheck": true,
     "resolveJsonModule": true,
-    "strict": true
+    "strict": true,
+    "paths": {
+      "@card/*": ["./src/modules/card/*"],
+      "@topic/*": ["./src/modules/topic/*"],
+      "@srs/*": ["./src/modules/srs/*"],
+      "@core/*": ["./src/core/*"],
+      "@pages/*": ["./src/pages/*"],
+      "@navigation/*": ["./src/navigation/*"],
+      "@root/*": ["./*"]
+    }
   }
 }
